v0.2.1-beta
Left lockout timer marked as beta but in fact have not been able to reproduce any issues with it when timer is set to 5 minutes or less. Issues only occurred with 15 and 30 mins, possibly related to how device sleeps, will continue looking into that and mark as beta for now even though options for 15 and 30 have been removed.

Fixed logical operator typos in QML that meant entries weren't getting necessary formatting.

Fixed Ko-fi link that was still pointing to previously used site.